+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Description: Robotiq 2F-85 v4 (URDF) (thanks to @nickswalker)
+
 ## [1.23.0] - 2026-03-09
 
 ### Added